[Context] Promoted commands as services
In order to put some context, we need to access the configuration and
some services in the Command configure method, so we can set default
values for options.

The configure method being called in the Command constructor, we don't
have access yet to the container (which is set afterward), which means
we need to promote commands as services and inject those dependencies
into the constructor.

As a side effect, the tests are no longer coupled to the database: we
pass a test double of repositories.