[Refactoring3] Created TaskCollection for activity
The display of tasks have the following logic:

* group tasks by a given index (day, project, title)
* compute the time spent in a given unit (working days, working hours)

This can be centralized into an object: TaskCollection.

You give it a TimeCard, and it returns a Task. If it doesn't have the
task, it creates it from the TimeCard's data.

This allows you to then add TimeCards to the task, so it will be able to
compute the time spent.

Timesheets will have the responsibility to retrieve TimeCard from the
database, and to return TaskCollections.